Prefer active Python on Poetry 1.2+
Set experimental virtualenvs.prefer-active-python config option, newly
introduced in Poetry 1.2.0, to true globally on Poetry 1.2+. Instruct
Poetry to use the active version of Python rather than the version of
Python that was active at the time the virtualenv was created. Spare the
end developer the hassle of running poetry env use
"$(asdf which python)" or blowing away and recreating their virtualenv
when swapping Python versions.